More mirror options
This commit allows parts of the mirror to be switched on and off and
paves the way for the mirror role to provide more services.
Additional services in the future may include Tor or alternate transit
schemes.  Additional options have been added to support secure rsyncd
configurations based on a list of hashes describing users, and a
secret hash containing the user passwords.