Remove archive-mirror infrastructure
The archive mirror hasn't been operational for 3 years now and if it
was to be brought back up, better engineering would need to be done to
make it work reliably.  This commit removes the variables and role
that were associated with the archive mirror to make things a little
cleaner to manage.