ansible: Improve Nomad reliability
This commit improves the reliability of Nomad on machines that have
just restarted.  It incorporates a suggestion from @ericonr to change
the unbound config to permit consul to come up independently of
unbound, and a change to the Nomad service to ensure that all of
Nomad's dependencies are available when starting.

One major issue remains, that is tracked as hashicorp/nomad#11272
wherein its possible for system tasks that have static port
allocations to fail to reschedule if there's dangling state on a node
after a reboot.  This is likely going to require a patch in Nomad
itself to resolve.