terraform/do: Provision f-sfo3-us.m.voidlinux.org.
This commit provisions f-sfo3-us.m.voidlinux.org which is intended to
be used as our new mail server.  As a result it is assigned a reserved
IP in addition to its ephemeral IP so that we can make use of the
reserved IP for reputation purposes.

In order to ensure that the new machine has correct rDNS, and to just
clean up the others at the same time, this commit also adjusts the
naming convention for VMs to be the FQDN instead of just the short
name.