services/nomad/build/buildbot: collect diff better to avoid skipping new packages
for some reason, when collapsing build requests, `self.build.allFiles()`
only includes the changes from the `HEAD` commit, which is especially
egregious when we jump to the `HEAD` commit (not the one that's supposed
to be built according to buildbot). By doing the diff ourselves, we can
avoid that pain and get the true list of packages that have changed from
the previous build to now.

This will still miss builds if there was a failure, though. We'd need a
better build orchestrator to fix that, I think.